requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.4.1
attrs==25.1.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
fonttools==4.55.3
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.12.2
groq==0.29.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.1
hyperframe==6.1.0
idna==3.10
imageio==2.36.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
narwhals==1.28.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.1
openai==1.93.0
packaging==24.1
pamqp==3.3.0
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.6.0
postgrest==1.1.1
prompt_toolkit==3.0.48
propcache==0.3.2
protobuf==5.29.3
psutil==6.1.0
pure_eval==0.2.3
pyarrow==19.0.1
pydantic==2.11.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.1
pypdf==5.6.0
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
realtime==2.5.3
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
rich==13.9.4
rich-toolkit==0.14.7
rpds-py==0.23.1
safetensors==0.5.3
scikit-image==0.25.0
scikit-learn==1.7.0
scipy==1.15.1
setuptools==80.9.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
storage3==0.12.0
streamlit==1.42.2
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
sympy==1.14.0
tenacity==9.0.0
threadpoolctl==3.6.0
tifffile==2025.1.10
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.53.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.3
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
